coldata: change Batch interface to operate with int length
This commit changes Batch interface to operate with int length
because it unifies the way we store batches - now we will have a single
implementation MemBatch (with an exception for coldata.ZeroBatch). This
allows us to remove bufferedBatch.

Release note: None

+++ pkg/col/coldata/random_testutils.go
@@ -121,7 +121,7 @@ func RandomVec(
 
 	for i := 0; i < n; i++ {
 		if rng.Float64() < nullProbability {
-			vec.Nulls().SetNull(uint16(i))
+			vec.Nulls().SetNull(i)
 		}
 	}
 }
